// src/internal/model/fileBucketResponse.ts
/**
 * Juno
 * Juno Public API Docs
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class FileBucketResponse {
  /**
   * The unique name of the registered bucket
   */
  'name': string;
  /**
   * Configuration ID associated with the registered bucket
   */
  'configId': number;
  /**
   * The name of the file provider associated with the bucket
   */
  'fileProviderName': string;
  /**
   * The list of file identifiers associated with the bucket
   */
  'fileServiceFile': Array<object>;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: 'name',
      baseName: 'name',
      type: 'string',
    },
    {
      name: 'configId',
      baseName: 'configId',
      type: 'number',
    },
    {
      name: 'fileProviderName',
      baseName: 'fileProviderName',
      type: 'string',
    },
    {
      name: 'fileServiceFile',
      baseName: 'FileServiceFile',
      type: 'Array<object>',
    },
  ];

  static getAttributeTypeMap() {
    return FileBucketResponse.attributeTypeMap;
  }
}