// src/internal/model/issueJWTResponse.ts
/**
 * Juno
 * Juno Public API Docs
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export class IssueJWTResponse {
  /**
   * Created JWT token
   */
  'token': string;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: 'token',
      baseName: 'token',
      type: 'string',
    },
  ];

  static getAttributeTypeMap() {
    return IssueJWTResponse.attributeTypeMap;
  }
}